Move query monads to `GeniusYield.TxBuilder.Query.Class`
